Calculation, Restructure

+++ seeder.js
@@ -4,6 +4,7 @@ const dotenv = require("dotenv");
 
 const User = require("./models/User");
 const Transaction = require("./models/Transaction");
+const Notification = require("./models/Notification");
 
 dotenv.config({ path: "./config/config.env" });
 
@@ -15,8 +16,9 @@ mongoose.connect(process.env.MONGO_URI, {
 
 const resetAccounts = async () => {
   try {
-    await User.updateMany({}, { balance: 100 });
+    await User.updateMany({}, { balance: 1000 });
     await Transaction.deleteMany();
+    await Notification.deleteMany();
     console.log("ACCOUNTS RESET!".yellow.inverse);
     process.exit();
   } catch (err) {
@@ -26,8 +28,9 @@ const resetAccounts = async () => {
 
 const resetDB = async () => {
   try {
-    await User.deleteMany();
+    await User.deleteMany({ role: { $ne: "admin" } });
     await Transaction.deleteMany();
+    await Notification.deleteMany();
     console.log("DB RESET!".red.inverse);
     process.exit();
   } catch (err) {
